Fix: Set Y axis minimum to zero
The axis range is set by the first histogram drawn on the canvas. For the plots containing the total, proportional, and pure rate this could mean that the much larger total rate causes the other rates to not appear in the plot.

Setting the Y axis to zero solves this issue.