Add logging for missing seeds
If the unprescaled seeds defined by the user are not present in the root file then the code will, instead of exiting, output the seeds in question at the start of the code and continue. The created histograms will still contain the missing seeds as labels but they won't be filled.